refactor: move EvmChainId and related constants to a new chains.ts file for better organization
