+++ frontend/components/Pages/AddBackupWalletViaSafePage/index.tsx
@@ -1,11 +1,11 @@
 import { Card, Flex, Skeleton, Typography } from 'antd';
 
 import { CardTitle } from '@/components/Card/CardTitle';
+import { AllEvmChainId, EvmChainId } from '@/constants/chains';
 import { UNICODE_SYMBOLS } from '@/constants/symbols';
 import { DISCORD_TICKET_URL } from '@/constants/urls';
 import { useServices } from '@/hooks/useServices';
 import { useMasterWalletContext } from '@/hooks/useWallet';
-import { AllEvmChainId, EvmChainId } from '@/types/Chain';
 
 import { GoToMainPageButton } from '../GoToMainPageButton';
